requirements.txt
docker==7.1.0
requests==2.32.4
python-dotenv==1.0.0
psutil==5.9.8
schedule==1.2.0 